refactor: ExamCard 컴포넌트 수정
- 현재 상태 및 뱃지를 나타내는 컴포넌트 분리
- 쪽지시험 아이콘 컴포넌트 분리
